requirements.txt
discord.py
python-dotenv
os
beautifulsoup4
requests